Move AIMS lib to lib/python
 * rename lib to aims_realtime_utils.py
 * modify SOOP TRV, FAIMMS and ANMN NRS to handle the new name
 * remove lib/aims from PYTHONPATH env
