+++ ANMN/NRS_AIMS/REALTIME/anmn_nrs_aims.py
@@ -35,17 +35,17 @@ from netCDF4 import Dataset
 from tendo import singleton
 
 from dest_path import get_anmn_nrs_site_name, get_main_anmn_nrs_var
-from realtime_util import (close_logger, convert_time_cf_to_imos,
-                           create_list_of_dates_to_download, download_channel,
-                           fix_data_code_from_filename,
-                           fix_provider_code_from_filename, get_channel_info,
-                           has_var_only_fill_value, is_above_file_limit,
-                           is_no_data_found, is_time_monotonic,
-                           is_time_var_empty, logging_aims, md5,
-                           modify_aims_netcdf, parse_aims_xml,
-                           remove_dimension_from_netcdf,
-                           remove_end_date_from_filename, save_channel_info,
-                           set_up)
+from aims_realtime_util import (close_logger, convert_time_cf_to_imos,
+                                create_list_of_dates_to_download, download_channel,
+                                fix_data_code_from_filename,
+                                fix_provider_code_from_filename, get_channel_info,
+                                has_var_only_fill_value, is_above_file_limit,
+                                is_no_data_found, is_time_monotonic,
+                                is_time_var_empty, logging_aims, md5,
+                                modify_aims_netcdf, parse_aims_xml,
+                                remove_dimension_from_netcdf,
+                                remove_end_date_from_filename, save_channel_info,
+                                set_up)
 from util import pass_netcdf_checker
